tstorbe/dcrdata: Cleanup plugin and add docs.
