+++ politeiawww/ticketvote/error.go
@@ -20,7 +20,7 @@ import (
 func respondWithError(w http.ResponseWriter, r *http.Request, format string, err error) {
 	var (
 		ue  v1.UserErrorReply
-		pde pdclient.Error
+		pde pdclient.RespError
 	)
 	switch {
 	case errors.As(err, &ue):
@@ -58,7 +58,7 @@ func respondWithError(w http.ResponseWriter, r *http.Request, format string, err
 	}
 }
 
-func handlePDError(w http.ResponseWriter, r *http.Request, format string, pde pdclient.Error) {
+func handlePDError(w http.ResponseWriter, r *http.Request, format string, pde pdclient.RespError) {
 	var (
 		pluginID   = pde.ErrorReply.PluginID
 		errCode    = pde.ErrorReply.ErrorCode
